# pkg/trace/{agent,api,filters}: refactor the agent Process function