--- a/pkg/serverless/invocationlifecycle/trace.go
+++ b/pkg/serverless/invocationlifecycle/trace.go
@@ -132,7 +132,7 @@ func endExecutionSpan(executionContext *ExecutionStartInfo, triggerTags map[stri
 	}
 
 	processTrace(&api.Payload{
-		Meta: api.Metadata{
+		Metadata: api.Metadata{
 			Source: info.NewReceiverStats().GetTagStats(info.Tags{}),
 		},
 		TracerPayload: tracerPayload,