--- a/pkg/serverless/trace/cold_start_span_creator.go
+++ b/pkg/serverless/trace/cold_start_span_creator.go
@@ -86,7 +86,9 @@ func (c *ColdStartSpanCreator) processSpan(coldStartSpan *pb.Span) {
 	}
 
 	c.traceAgent.ta.Process(&api.Payload{
-		Source:        info.NewReceiverStats().GetTagStats(info.Tags{}),
+		Meta: api.Metadata{
+			Source: info.NewReceiverStats().GetTagStats(info.Tags{}),
+		},
 		TracerPayload: tracerPayload,
 	})
 }