--- a/pkg/serverless/trace/cold_start_span_creator.go
+++ b/pkg/serverless/trace/cold_start_span_creator.go
@@ -86,7 +86,7 @@ func (c *ColdStartSpanCreator) processSpan(coldStartSpan *pb.Span) {
 	}
 
 	c.traceAgent.ta.Process(&api.Payload{
-		Meta: api.Metadata{
+		Metadata: api.Metadata{
 			Source: info.NewReceiverStats().GetTagStats(info.Tags{}),
 		},
 		TracerPayload: tracerPayload,