--- a/pkg/serverless/trace/inferredspan/inferred_span.go
+++ b/pkg/serverless/trace/inferredspan/inferred_span.go
@@ -132,7 +132,7 @@ func (inferredSpan *InferredSpan) CompleteInferredSpan(
 	}
 
 	processTrace(&api.Payload{
-		Meta: api.Metadata{
+		Metadata: api.Metadata{
 			Source: info.NewReceiverStats().GetTagStats(info.Tags{}),
 		},
 		TracerPayload: tracerPayload,