--- a/pkg/trace/agent/agent.go
+++ b/pkg/trace/agent/agent.go
@@ -81,7 +81,7 @@ type Agent struct {
 	ctx context.Context
 
 	// tfs is the list of transformers used by agent.Process()
-	tfs []Transformer
+	transformers []Transformer
 }
 
 // NewAgent returns a new Agent object, ready to be started. It takes a context
@@ -116,6 +116,7 @@ func NewAgent(ctx context.Context, conf *config.AgentConfig) *Agent {
 	agnt.Receiver = api.NewHTTPReceiver(conf, dynConf, in, agnt)
 	agnt.OTLPReceiver = api.NewOTLPReceiver(in, conf)
 	agnt.RemoteConfigHandler = remoteconfighandler.New(conf, agnt.PrioritySampler, agnt.RareSampler, agnt.ErrorsSampler)
+	agnt.transformers = agnt.makeTransformers()
 	return agnt
 }
 
@@ -328,41 +329,40 @@ func computeTopLevel(m *api.Metadata, pt *traceutil.ProcessedTrace) error {
 	return nil
 }
 
-func (a *Agent) transformers() []Transformer {
-	if a.tfs == nil {
-		a.tfs = []Transformer{
-			normalizePayloadEnv,
-			a.discardSpans,
-			rejectEmpty,
-			countSpansReceived,
-			normalizeChunk,
-			normalizeSpans,
-			a.blocklist,
-			a.tagFilter,
-			eachSpan(
-				a.applyGlobalTags,
-				a.modifySpan,
-				a.obfuscateSpan,
-				Truncate,
-				updateTopLevel,
-				a.Replacer.Replace,
-			),
-			a.setRootSpanTags,
-			computeTopLevel,
-		}
+func (a *Agent) makeTransformers() []Transformer {
+	return []Transformer{
+		normalizePayloadEnv,
+		a.discardSpans,
+		rejectEmpty,
+		countSpansReceived,
+		normalizeChunk,
+		normalizeSpans,
+		a.blocklist,
+		a.tagFilter,
+		eachSpan(
+			a.applyGlobalTags,
+			a.modifySpan,
+			a.obfuscateSpan,
+			Truncate,
+			updateTopLevel,
+			a.Replacer.Replace,
+		),
+		a.setRootSpanTags,
+		computeTopLevel,
 	}
-	return a.tfs
 }
 
-// Transformer is a transformation of a processed trace.
+// Transformer is a transformation of a processed trace. If a Transformer returns an error, the
+// ProcessedTrace will be discarded and not sent to the intake.
 type Transformer func(*api.Metadata, *traceutil.ProcessedTrace) error
 
-// SpanTransformer is a transformation of a span within a trace.
+// SpanTransformer is a transformation of a span within a trace. If a SpanTransformer returns an
+// error, the ProcessedTrace will be discarded and not sent to the intake.
 type SpanTransformer func(*api.Metadata, *traceutil.ProcessedTrace, *pb.Span) error
 
 // transform applies all of the transformations in a.transformers
 func (a *Agent) transform(m *api.Metadata, pt *traceutil.ProcessedTrace) error {
-	for _, t := range a.transformers() {
+	for _, t := range a.transformers {
 		err := t(m, pt)
 		if err != nil {
 			return err