… pkg/serverless/trace/span_modifer_test.go
… pkg/trace/agent/agent_test.go
… pkg/trace/agent/fuzz_test.go
… pkg/trace/agent/normalizer_test.go
… pkg/trace/agent/obfuscate.go
… pkg/trace/agent/obfuscate_test.go
… pkg/trace/agent/truncator.go
… pkg/trace/agent/truncator_test.go
… pkg/trace/api/api.go
… pkg/trace/api/api_test.go
… pkg/trace/api/otlp.go
… pkg/trace/api/otlp_test.go
… pkg/trace/api/payload.go
… pkg/trace/filters/replacer.go
… pkg/trace/filters/replacer_test.go