@knusbaum 2022-12-01
Benchmarks don't reveal any difference between the old processing step and the new one:
```
$ benchstat main.bench.out processing.bench.out 
name                                           old time/op    new time/op    delta
AgentTraceProcessing-16                          186µs ±127%     200µs ±90%    ~     (p=0.579 n=10+10)
AgentTraceProcessingWithFiltering-16             88.6µs ± 9%   146.2µs ±98%    ~     (p=0.274 n=8+10)
AgentTraceProcessingWithWorstCaseFiltering-16     165µs ±45%     149µs ±54%    ~     (p=0.315 n=9+10)

name                                           old alloc/op   new alloc/op   delta
AgentTraceProcessing-16                          80.0kB ± 1%    81.1kB ± 1%  +1.35%  (p=0.000 n=10+8)
AgentTraceProcessingWithFiltering-16             69.8kB ± 1%    71.6kB ± 1%  +2.64%  (p=0.000 n=10+10)
AgentTraceProcessingWithWorstCaseFiltering-16    79.8kB ± 1%    80.5kB ± 2%  +0.97%  (p=0.023 n=10+10)

name                                           old allocs/op  new allocs/op  delta
AgentTraceProcessing-16                             784 ± 1%       788 ± 2%    ~     (p=0.233 n=9+10)
AgentTraceProcessingWithFiltering-16                762 ± 1%       767 ± 1%  +0.78%  (p=0.001 n=10+9)
AgentTraceProcessingWithWorstCaseFiltering-16       791 ± 2%       783 ± 2%  -1.05%  (p=0.027 n=10+10)